regroup auth calls
