+++ src/composables/useOAuth.ts
@@ -1,0 +1,28 @@
+import { LocalStorageKeys } from "@/store/lskeys";
+import { User } from "@/structures/User";
+
+export function useOAuth() {
+	function prompt(provider: User.Connection.Platform, token?: string | null): Promise<string> {
+		const w = window.open(
+			`${import.meta.env.VITE_APP_API_REST}/auth?platform=${provider.toLowerCase()}` +
+				(token ? `&token=${token}` : ""),
+			"seventv-oauth2",
+			"_blank, width=850, height=650, menubar=no, location=no",
+		);
+
+		return new Promise((resolve) => {
+			const i = setInterval(async () => {
+				if (!w?.closed) {
+					return;
+				}
+
+				clearInterval(i);
+				resolve(localStorage.getItem(LocalStorageKeys.TOKEN) || "");
+			}, 100);
+		});
+	}
+
+	return {
+		prompt,
+	};
+}
