+++ src/views/user/UserDetails.vue
@@ -119,6 +119,7 @@ import { Permissions } from "@/structures/Role";
 import { User } from "@/structures/User";
 import { ConvertIntColorToHex } from "@/structures/util/Color";
 import { useContext } from "@/composables/useContext";
+import { useOAuth } from "@/composables/useOAuth";
 import Icon from "@/components/utility/Icon.vue";
 import UserTag from "@/components/utility/UserTag.vue";
 import UserEditorModal from "./UserEditorModal.vue";
@@ -129,6 +130,7 @@ const UserRoleList = defineAsyncComponent(() => import("@/components/utility/Use
 const { t } = useI18n();
 
 const actor = useActor();
+const oauth = useOAuth();
 
 const ctx = useContext("USER");
 if (!ctx) throw new Error("No user context provided");
@@ -163,13 +165,7 @@ const linkAccount = (platform: User.Connection.Platform) => {
 		return;
 	}
 
-	window.open(
-		`${import.meta.env.VITE_APP_API_REST}/auth/${platform.toLowerCase()}?token=${localStorage.getItem(
-			LocalStorageKeys.TOKEN,
-		)}`,
-		"7TVOAuth2",
-		"_blank, width=850, height=650, menubar=no, location=no",
-	);
+	oauth.prompt(platform, localStorage.getItem(LocalStorageKeys.TOKEN));
 };
 
 const modal = useModal();
